+++ uikit-samples/src/main/java/com/sendbird/uikit/samples/basic/openchannel/livestream/LiveStreamActivity.kt
@@ -79,7 +79,7 @@ class LiveStreamActivity : AppCompatActivity() {
             }
             addChannelHandler()
             ivLive.visibility = View.VISIBLE
-            ivLive.setOnClickListener {groupLiveControl.toggleVisibility() }
+            ivLive.setOnClickListener { groupLiveControl.toggleVisibility() }
             ivClose.setOnClickListener { finish() }
             ivChatToggle?.setOnClickListener {
                 if (sbFragmentContainer.visibility == View.GONE) {
@@ -189,11 +189,11 @@ class LiveStreamActivity : AppCompatActivity() {
     private fun hideSystemUI() {
         if (resources.configuration.orientation == Configuration.ORIENTATION_LANDSCAPE) {
             window.decorView.systemUiVisibility = View.SYSTEM_UI_FLAG_LAYOUT_STABLE or
-                    View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION or
-                    View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN or
-                    View.SYSTEM_UI_FLAG_HIDE_NAVIGATION or
-                    View.SYSTEM_UI_FLAG_FULLSCREEN or
-                    View.SYSTEM_UI_FLAG_LOW_PROFILE or View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY
+                View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION or
+                View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN or
+                View.SYSTEM_UI_FLAG_HIDE_NAVIGATION or
+                View.SYSTEM_UI_FLAG_FULLSCREEN or
+                View.SYSTEM_UI_FLAG_LOW_PROFILE or View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY
         }
     }