+++ uikit/src/main/java/com/sendbird/uikit/internal/interfaces/OnImpressionDetectedListener.kt
@@ -1,0 +1,9 @@
+package com.sendbird.uikit.internal.interfaces
+
+/**
+ * On visible item detect listener
+ */
+@JvmSuppressWildcards
+internal fun interface OnImpressionDetectedListener<T> {
+    fun onImpressionDetected(items: List<T>)
+}